refactor(PE-1271): consolidate multi-decoder variant into MULTI_DECODER flag
