--- examples/browser/seamless-video-switching-multi-decoder/index.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Seamless Video App</title>
-    <style>
-      body,
-      html {
-        margin: 0;
-        padding: 0;
-        height: 100%;
-        overflow: hidden;
-      }
-
-      body {
-        font-family: Arial, sans-serif;
-        color: white;
-      }
-
-      #video-container {
-        width: 100%;
-        height: 100%;
-      }
-
-      .hidden {
-        display: none;
-      }
-
-      #video-player-1,
-      #video-player-2 {
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-      }
-
-      video {
-        background: black;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="video-container">
-      <video id="video-player-1" preload="auto" autoplay></video>
-      <video id="video-player-2" class="hidden" preload="auto"></video>
-    </div>
-    <script src="index.js"></script>
-  </body>
-</html>
